Replace stored credentials with persona test user accounts
