--- credentials.yaml
@@ -1,38 +1,0 @@
-# This Source Code Form is subject to the terms of the Mozilla Public
-# License, v. 2.0. If a copy of the MPL was not distributed with this
-# file, You can obtain one at http://mozilla.org/MPL/2.0/.
-#
-#
-# File contains users data.
-#
-# Each user is a section named with its role
-# and any number of values. At least email,
-# password and name should be present.
-#
-# Example:
-#     admin:
-#         email: email@site.com
-#         password: password
-#         name: Test User
-#
-# Still, you are free to add any more data you wish. It will be kept
-# in the same dictionary.
-#
-# Example:
-#     admin:
-#         email: email@site.com
-#         password: password
-#         name: Test User
-#         username: testuser
-#         some_user_data: data
-#
-# The contents of this file are accessible via the pytest-mozwebqa plugin:
-#
-# Example:
-#   credentials = mozwebqa.credentials['default']
-#   credentials['email']
-
-default:
-    email: <value>
-    password: <value>
-    name: <value>
