+++ pages/base.py
@@ -17,13 +17,13 @@ from pages.page import Page
 class Base(Page):
 
     _page_title_locator = (By.CSS_SELECTOR, 'h1.page-title')
+    _page_header_locator = (By.CSS_SELECTOR, '.contain > h1')
 
-    #LoggedIn
+    _login_browser_id_locator = (By.CSS_SELECTOR, 'a.persona-button')
     _logout_locator = (By.CSS_SELECTOR, '#nav-user-submenu li a.browserid-logout')
     _profile_locator = (By.CSS_SELECTOR, '#nav-user-submenu li:nth-of-type(2) a')
     _username_locator = (By.CSS_SELECTOR, '#nav-main-menu > li.user > a')
 
-    #Content Navigation
     _about_content_nav_locator = (By.CSS_SELECTOR, '#nav-main-menu li:nth-of-type(1) a')
     _leaderboard_link_locator = (By.CSS_SELECTOR, '#nav-main-menu > li:nth-of-type(2) a')
 
@@ -43,16 +43,29 @@ class Base(Page):
     def username(self):
         return self.selenium.find_element(*self._username_locator).text
 
-    def get_new_persona_credentials(self):
-        url = "http://personatestuser.org/email/"
-        response = urllib2.urlopen(url).read()
-        decode = json.loads(response)
-        credentials = {
-            'email': decode['email'],
-            'password': decode['pass']
-        }
+    def _create_persona_test_user(self):
+        response = urllib2.urlopen('http://personatestuser.org/email/').read()
+        user = json.loads(response)
+        return user['email'], user['pass']
 
-        return credentials
+    def login(self, email=None, password=None):
+        if not all([email, password]):
+            email, password = self._create_persona_test_user()
+        self.click_login()
+        from browserid import BrowserID
+        pop_up = BrowserID(self.selenium, self.timeout)
+        pop_up.sign_in(email, password)
+        WebDriverWait(self.selenium, self.timeout).until(
+            lambda s: self.is_user_logged_in)
+        from pages.home import Home
+        return Home(self.testsetup)
+
+    def click_login(self):
+        self.selenium.find_element(*self._login_browser_id_locator).click()
+
+    @property
+    def header(self):
+        return self.selenium.find_element(*self._page_header_locator).text.replace('\n', ' ')
 
     def logout(self):
         self._hover_user_menu()
@@ -72,13 +85,6 @@ class Base(Page):
         self.selenium.find_element(*self._about_content_nav_locator).click()
         from pages.about import About
         return About(self.testsetup)
-
-    def login_user_browser_id(self, user):
-        credentials = self.testsetup.credentials[user]
-        from browserid import BrowserID
-        browserid = BrowserID(self.selenium, self.timeout)
-        browserid.sign_in(credentials['email'], credentials['password'])
-        WebDriverWait(self.selenium, self.timeout).until(lambda s: s.find_element(*self._logout_locator))
 
     def click_leaderboard_link(self):
         self.selenium.find_element(*self._leaderboard_link_locator).click()
