+++ requirements.txt
@@ -1,4 +1,3 @@
-PyYAML==3.10
 UnittestZero
 execnet==1.0.9
 py==1.4.8
