+++ tests/test_about_page.py
@@ -9,18 +9,15 @@ import pytest
 from pages.start_page import StartPage
 from unittestzero import Assert
 
-credentials = pytest.mark.credentials
 nondestructive = pytest.mark.nondestructive
 
 
 class TestAboutPage:
 
-    @credentials
     @nondestructive
     def test_about_page_has_proper_layout(self, mozwebqa):
         start_page = StartPage(mozwebqa)
         home_page = start_page.login()
-
         about_page = home_page.click_about_nav_link()
         Assert.true(about_page.is_the_current_url)
         Assert.equal(about_page.faq_header, 'Frequently Asked Questions',
